requirements.txt
huggingface_hub[cli]
accelerate==1.12.0
transformers
torch

ffmpeg-python
flask
azure-storage-blob

pytest
pytest-cov

PyYAML
python-dotenv

soundfile
azure-ai-translation-text
azure-ai-transcription
qwen-tts
kafka-python
psycopg[binary]
